Move some functions into a base entry handler class
Reduces duplication of code
