+++ src/handler/entry_handler/EntryHandler.py
@@ -1,11 +1,11 @@
 import pandas
 
-from src.csv.CSVAttributes import CSVAttributes
 from src.exceptions.InvalidIDOfDateException import InvalidIDOfDateException
+from src.handler.entry_handler.BaseEntryHandlerClass import BaseEntryHandlerClass
 from src.handler.entry_handler.GroupedEntryHandler import GroupedEntryHandler
 
 
-class EntryHandler(CSVAttributes):
+class EntryHandler(BaseEntryHandlerClass):
     def __init__(self):
         super(EntryHandler, self).__init__()
 
@@ -19,26 +19,26 @@ class EntryHandler(CSVAttributes):
         except KeyError:
             raise InvalidIDOfDateException
 
-        self.grouped_entry_handler.data = self.grouped_entry_handler.get_data()
-        self.grouped_entry_handler.data = self.grouped_entry_handler.group_entries_by_date()
+        self.data = self.get_data()
+        self.data = self.group_entries_by_date()
 
-        entries_of_date_at_index: pandas.DataFrame = self.grouped_entry_handler.data.get_group(date)
-        entries_of_date_at_index.reset_index()
-        entries_of_date_at_index.index += 1
+        self.data: pandas.DataFrame = self.data.get_group(date)
+        self.data.reset_index()
+        self.boost_index()
 
-        entries_of_date_at_index = entries_of_date_at_index \
+        self.data = self.data \
             .rename(columns={
                                 "start_date": "Start Date", "start_time": "Start Time",
                                 "stop_date": "Stop Date", "stop_time": "Stop Time",
                                 "work_hours": "Work hours", "message": "Message"
                             }
                     )
-        entries_of_date_at_index["Start"] = entries_of_date_at_index[["Start Date", "Start Time"]].apply(lambda x: " at ".join(x), axis=1)
-        entries_of_date_at_index["Stop"] = entries_of_date_at_index[["Stop Date", "Stop Time"]].apply(lambda x: " at ".join(x), axis=1)
+        self.data["Start"] = self.data[["Start Date", "Start Time"]].apply(lambda x: " at ".join(x), axis=1)
+        self.data["Stop"] = self.data[["Stop Date", "Stop Time"]].apply(lambda x: " at ".join(x), axis=1)
 
-        entries_of_date_at_index = entries_of_date_at_index.drop(columns=["Start Date", "Start Time", "Stop Date", "Stop Time"])
-        entries_of_date_at_index = entries_of_date_at_index[["Start", "Stop", "Work hours", "Message"]]
+        self.data = self.data.drop(columns=["Start Date", "Start Time", "Stop Date", "Stop Time"])
+        self.data = self.data[["Start", "Stop", "Work hours", "Message"]]
 
         pandas.set_option("display.max_colwidth", None)
 
-        return [date, entries_of_date_at_index]
+        return [date, self.data]
